+++ src/modules/automata.rs
@@ -3,7 +3,7 @@ use std::fmt::{Debug, Formatter, Result};
 use std::rc::Rc;
 use std::str::Chars;
 
-use super::grammar::Anchor;
+use super::regex::Anchor;
 use super::state::{AnchorState, LambdaState, State, TokenState, TrivialState};
 
 type StatePtr = Rc<RefCell<dyn State>>;
