+++ src/modules/mod.rs
@@ -1,9 +1,4 @@
-mod alphabet;
-mod ast;
 mod automata;
 mod error;
-mod grammar;
-pub mod language;
-mod monadic_parser;
-pub mod regexp;
+pub mod regex;
 mod state;
