+++ src/modules/regex/language.rs
@@ -1,6 +1,6 @@
+use super::{Automata, Error};
+
 use super::ast::AbstractSyntaxTree;
-use super::automata::Automata;
-use super::error::Error;
 use super::grammar::{self, Grammar, Regex};
 
 /// A modal for formal language
